+++ CHANGELOG.md
@@ -38,6 +38,8 @@ All notable changes to this project will be documented in this file. This change
 * grid functions accept separate coordinates
 * many additional `log-exp` functions
 * weighted mean (`wmean`), median (`wmedian`) and quantiles (`wquantile`,`wquantiles`)
+* Vec2/3/4 implement IReduce and IReduceInit
+* historical trigonometric functions (versine with variants, chord, exsec/excsc) with their inverses
 
 #### Tests
 
@@ -72,6 +74,7 @@ All notable changes to this project will be documented in this file. This change
 
 * a nasty bug with primitive macro generation, how could I missed that?
 * `:histogram` issues with low number of samples
+* `fast-max` and `fast-min` had wrong inline operation
 
 ## [2.1.8]